refactor: isolate RadioGroup stories
